+++ OMG.PSUtilities.AI/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.AI
+
+## [1.0.0] - 2025-07-16
+- OMG.PSUtilities.AI.psd1 : Added functions : 'Invoke-PSUPromptAI' and 'Start-PSUAiChat'
+- OMG.PSUtilities.AI.psm1 : Added the code to load the private and public functions into the session, and further export public functions.