+++ OMG.PSUtilities.ActiveDirectory/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.ActiveDirectory
+
+## [1.0.0] - 2025-07-16
+- OMG.PSUtilities.ActiveDirectory.psd1 : Exported a dummy function for testing.
+- OMG.PSUtilities.ActiveDirectory.psm1 : Added the code to load the private and public functions into the session, and further export public functions.