+++ OMG.PSUtilities.AzureCore/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.AzureCore
+
+## [1.0.0] - 2025-07-16
+- OMG.PSUtilities.AzureCore.psd1 : Added functions 'Get-PSUAzToken' and  'Test-PSUAzConnection'
+- OMG.PSUtilities.AzureCore.psm1 : Added the code to load the private and public functions into the session, and further export public functions.