+++ OMG.PSUtilities.AzureDevOps/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.AzureDevOps
+
+## [1.0.0] - 2025-07-16
+- OMG.PSUtilities.AzureDevOps.psd1 : Added functions 'Get-PSUADOBuildDetails' and 'Get-PSUADOPipelineLatestRun'
+- OMG.PSUtilities.AzureDevOps.psm1 : Added the code to load the private and public functions into the session, and further export public functions.