+++ OMG.PSUtilities.ServiceNow/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.ServiceNow
+
+## [1.0.0] - 2025-07-16
+- OMG.PSUtilities.ServiceNow.psd1 : Added a dummy function (New-OMGPSUtilitiesServiceNow) for testing.
+- OMG.PSUtilities.ServiceNow.psm1 : Added the code to load the private and public functions into the session, and further export public functions.