+++ OMG.PSUtilities.VSphere/CHANGELOG.md
@@ -1,2 +1,6 @@
 ## Changelog
 - Initial scaffolding for OMG.PSUtilities.VSphere
+
+## [1.0.0] - 2025-07-16
+- OMG.PSUtilities.VSphere.psd1 : Added dummy function ('New-OMGPSUtilitiesVSphere') for testing.
+- OMG.PSUtilities.VSphere.psm1 : Added the code to load the private and public functions into the session, and further export public functions.